Reorganized the prefab folder and made a scene template
made a sub folder of the prefab folder containing all the stuff that you need in a level.

made a scene template called Basic Scene that you can use when creating a new scene to save you some time.

Instead of dragging in all of the stuff into a scene, you make a new scene from the Basic Scene scene template

If you want to update the template, update the Basic Scene scene, then save it as a Scene template and replace the old Basic Scene scene template