Updated goal flag partcles, doubled camera size and background size
I doubled the size of the camera and the background, so that there is now more room for levels. Designers should adjust the camera size to determine what gives them the most amount of space without making it too hard to see things. Doubling the camera size also broke every level so those will need to be fixed